### 2025-05-17 #8218 @arfon
Moving this forward without additional review because of https://github.com/ropensci/software-review/issues/684
