### 2025-05-17 #8218 @arfon
@rmgpanw - looks like the affiliations aren't quite right here. Please review https://joss.readthedocs.io/en/latest/example_paper.html and update your paper.
