### 2025-05-17 #8218 @rmgpanw
Thanks @arfon, I have now addressed this in commit [c56e678](https://github.com/ropensci/gtexr/commit/c56e678745baa027832f54a8978d181cd67e151c)
