### 2025-05-17 #8218 @arfon
@ernestguevarra, @zkamvar – just to be clear here, there's no additional work for you to do, we're just crediting you for your reviews in https://github.com/ropensci/software-review/issues/684
