### 2025-05-17 #8249 @editorialbot
[REVIEW]: gtexr: A convenient R interface to the Genotype-Tissue Expression (GTEx) Portal API
**Submitting author:** @rmgpanw (<a href="http://orcid.org/0000-0002-0800-2890">Alasdair Warwick</a>)
**Repository:** https://github.com/ropensci/gtexr
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
