### 2025-05-17 #8249 @rmgpanw
@arfon no, not yet. Should I make one at this stage? There are a couple of releases on GitHub (corresponding to the 2 versions on CRAN), but not zenodo
