+++ .github/workflows/pkgcheck.yaml
@@ -20,3 +20,5 @@ jobs:
       issues: write
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
+        with:
+          summary-only: false
