### 2025-05-28 b9421b1
update citation details with JOSS paper
